requirements.txt
asgiref==3.5.2
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1
django-cors-headers==3.13.0
django-import-export==3.0.1
django-jazzmin==2.6.0
django-rest-multiple-models==2.1.3
django-unfold==0.21.1
djangorestframework==3.13.1
et-xmlfile==1.1.0
gunicorn==20.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.10
Pillow==9.2.0
psycopg2-binary==2.9.6
pycodestyle==2.9.1
pytz==2022.2.1
PyYAML==6.0
sqlparse==0.4.2
tablib==3.2.1
toml==0.10.2
whitenoise==6.2.0
xlrd==2.0.1
xlwt==1.3.0
